refactor(common): CharacterDetailModal 리팩토링 및 전체 복잡도 개선 완료
- CharacterDetailModal.tsx: CC 20 -> <10, 라인 수 407 -> 67 감소
- Components: CharacterHeader, Traits, Arc, Relationships, Appearances, AdditionalDetails 분리
- Hooks: useCharacterData로 데이터 파싱 로직 분리
- 프로젝트 내 모든 CC >= 15 컴포넌트 리팩토링 완료
